+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Tar archive import support
+
 ## [1.3.1] - 2025-05-28
 ### Changed
 - Updated dependencies